+++ airbyte-server/src/test/java/io/airbyte/server/handlers/ArchiveHandlerTest.java
@@ -24,107 +24,28 @@
 
 package io.airbyte.server.handlers;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.google.common.collect.ImmutableMap;
 import io.airbyte.commons.io.FileTtlManager;
-import io.airbyte.commons.json.Jsons;
-import io.airbyte.commons.resources.MoreResources;
-import io.airbyte.config.DestinationConnection;
-import io.airbyte.config.SourceConnection;
-import io.airbyte.config.StandardDestinationDefinition;
-import io.airbyte.config.StandardSourceDefinition;
-import io.airbyte.config.StandardSync;
-import io.airbyte.config.StandardSyncSchedule;
-import io.airbyte.config.StandardWorkspace;
-import io.airbyte.config.persistence.ConfigNotFoundException;
 import io.airbyte.config.persistence.ConfigRepository;
-import io.airbyte.config.persistence.PersistenceConstants;
-import io.airbyte.db.Database;
-import io.airbyte.db.Databases;
-import io.airbyte.scheduler.persistence.DefaultJobPersistence;
 import io.airbyte.scheduler.persistence.JobPersistence;
-import io.airbyte.server.helpers.ConnectionHelpers;
-import io.airbyte.server.helpers.DestinationDefinitionHelpers;
-import io.airbyte.server.helpers.DestinationHelpers;
-import io.airbyte.server.helpers.SourceDefinitionHelpers;
-import io.airbyte.server.helpers.SourceHelpers;
 import io.airbyte.validation.json.JsonValidationException;
 import java.io.IOException;
-import java.sql.SQLException;
-import java.util.List;
-import java.util.UUID;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.PostgreSQLContainer;
 
 public class ArchiveHandlerTest {
 
   private ConfigRepository configRepository;
-
-  private JsonNode dbConfig;
-  private PostgreSQLContainer<?> container;
-  private Database database;
-  private JobPersistence persistence;
   private ArchiveHandler archiveHandler;
-  private FileTtlManager fileTtlManager;
 
   @BeforeEach
   void setUp() {
     configRepository = mock(ConfigRepository.class);
-    dbConfig = null;
-    container = mock(PostgreSQLContainer.class);
-    database = mock(Database.class);
-    persistence = mock(JobPersistence.class);
-    fileTtlManager = mock(FileTtlManager.class);
-    archiveHandler = new ArchiveHandler("test-version", configRepository, persistence, fileTtlManager);
-  }
-
-  void setUpDatabase() throws IOException, SQLException {
-    container = new PostgreSQLContainer<>("postgres:13-alpine");
-    container.start();
-    dbConfig = Jsons.jsonNode(ImmutableMap.builder()
-        .put("username", container.getUsername())
-        .put("password", container.getPassword())
-        .put("jdbc_url", String.format("jdbc:postgresql://%s:%s/%s",
-            container.getHost(),
-            container.getFirstMappedPort(),
-            container.getDatabaseName()))
-        .build());
-    database = Databases.createPostgresDatabase(
-        dbConfig.get("username").asText(),
-        dbConfig.get("password").asText(),
-        dbConfig.get("jdbc_url").asText());
-    persistence = new DefaultJobPersistence(database);
-    final String schemaFile = MoreResources.readResource("schema.sql");
-    final String sql = schemaFile.substring(schemaFile.indexOf("-- Statements Below"), schemaFile.indexOf("-- Statements Above"));
-    database.query(ctx -> ctx.execute(sql));
-  }
-
-  @AfterEach
-  void tearDown() throws Exception {
-    database.close();
-    container.close();
-  }
-
-  private StandardWorkspace generateWorkspace() {
-    final UUID workspaceId = PersistenceConstants.DEFAULT_WORKSPACE_ID;
-
-    return new StandardWorkspace()
-        .withWorkspaceId(workspaceId)
-        .withCustomerId(UUID.randomUUID())
-        .withEmail("test@airbyte.io")
-        .withName("test workspace")
-        .withSlug("default")
-        .withInitialSetupComplete(false)
-        .withOnboardingComplete(true);
+    archiveHandler = new ArchiveHandler("test-version", configRepository, mock(JobPersistence.class), mock(FileTtlManager.class));
   }
 
   @Test
@@ -138,69 +59,6 @@ public class ArchiveHandlerTest {
     verify(configRepository, never()).writeDestinationConnection(any());
     verify(configRepository, never()).writeStandardSync(any());
     verify(configRepository, never()).writeStandardSchedule(any());
-  }
-
-  @Test
-  void testConfigMigration() throws ConfigNotFoundException, IOException, JsonValidationException {
-    final StandardWorkspace workspace = generateWorkspace();
-    final StandardSourceDefinition standardSource = SourceDefinitionHelpers.generateSource();
-    final StandardDestinationDefinition standardDestination = DestinationDefinitionHelpers.generateDestination();
-    final SourceConnection sourceConnection1 = SourceHelpers.generateSource(standardSource.getSourceDefinitionId());
-    final SourceConnection sourceConnection2 = SourceHelpers.generateSource(standardSource.getSourceDefinitionId());
-    final DestinationConnection destinationConnection = DestinationHelpers.generateDestination(standardDestination.getDestinationDefinitionId());
-    final StandardSync destinationSync = ConnectionHelpers.generateSyncWithDestinationId(destinationConnection.getDestinationId());
-    final StandardSync sourceSync = ConnectionHelpers.generateSyncWithSourceId(sourceConnection1.getSourceId());
-    final StandardSyncSchedule syncSchedule = ConnectionHelpers.generateSchedule(sourceSync.getConnectionId());
-
-    // Read operations
-    when(configRepository.getStandardWorkspace(PersistenceConstants.DEFAULT_WORKSPACE_ID)).thenReturn(workspace);
-    when(configRepository.listStandardSources()).thenReturn(List.of(standardSource));
-    when(configRepository.listStandardDestinationDefinitions()).thenReturn(List.of(standardDestination));
-    when(configRepository.listSourceConnection()).thenReturn(List.of(sourceConnection1, sourceConnection2));
-    when(configRepository.listDestinationConnection()).thenReturn(List.of(destinationConnection));
-    when(configRepository.listStandardSyncs()).thenReturn(List.of(destinationSync, sourceSync));
-    when(configRepository.getStandardSyncSchedule(sourceSync.getConnectionId())).thenReturn(syncSchedule);
-
-    archiveHandler.importData(archiveHandler.exportData());
-
-    verify(configRepository).writeStandardWorkspace(workspace);
-    verify(configRepository).writeStandardSource(standardSource);
-    verify(configRepository).writeStandardDestinationDefinition(standardDestination);
-    verify(configRepository).writeSourceConnection(sourceConnection1);
-    verify(configRepository).writeSourceConnection(sourceConnection2);
-    verify(configRepository).writeDestinationConnection(destinationConnection);
-    verify(configRepository).writeStandardSync(sourceSync);
-    verify(configRepository).writeStandardSync(destinationSync);
-    verify(configRepository).writeStandardSchedule(syncSchedule);
-    verify(fileTtlManager).register(any());
-  }
-
-  @Test
-  void testWrongDatabaseMigration() throws SQLException, IOException {
-    setUpDatabase();
-    database.query(ctx -> {
-      ctx.fetch("CREATE TABLE id_and_name(id INTEGER, name VARCHAR(200), updated_at DATE);");
-      ctx.fetch(
-          "INSERT INTO id_and_name (id, name, updated_at) VALUES (1,'picard', '2004-10-19'),  (2, 'crusher', '2005-10-19'), (3, 'vash', '2006-10-19');");
-      return null;
-    });
-    archiveHandler = new ArchiveHandler("test-version", configRepository, persistence, fileTtlManager);
-    assertThrows(RuntimeException.class, () -> archiveHandler.importData(archiveHandler.exportData()));
-  }
-
-  @Test
-  void testDatabaseMigration() throws SQLException, IOException {
-    setUpDatabase();
-    database.query(ctx -> {
-      ctx.fetch(
-          "INSERT INTO jobs(id, scope, config_type, config, status, created_at, started_at, updated_at) VALUES "
-              + "(1,'get_spec_scope', 'get_spec', '{ \"type\" : \"getSpec\" }', 'succeeded', '2004-10-19', null, '2004-10-19'), "
-              + "(2,'sync_scope', 'sync', '{ \"job\" : \"sync\" }', 'running', '2005-10-19', null, '2005-10-19'), "
-              + "(3,'sync_scope', 'sync', '{ \"job\" : \"sync\" }', 'pending', '2006-10-19', null, '2006-10-19');");
-      return null;
-    });
-    archiveHandler = new ArchiveHandler("test-version", configRepository, persistence, fileTtlManager);
-    archiveHandler.importData(archiveHandler.exportData());
   }
 
 }
